requirements.txt
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
colorlog==6.10.1
croniter==6.0.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Markdown==3.9
mutagen==1.47.0
pdfkit==1.0.0
pydub==0.25.1
pylast==6.0.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
schedule==1.2.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
tqdm==4.67.1
urllib3==2.5.0
wheel==0.46.1
pytest==8.4.2
pytest-mock==3.15.1
pytest-cov==7.0.0
pytest-asyncio==1.2.0
openai>=1.0.0
tiktoken
aiohttp>=3.9.0
PyYAML>=6.0
python-dotenv>=1.0.0
